+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.0.3 - 2022-04-11
+
+### Added
+- Added price splitting and merging functionality in examples/
+- Added support for fees in a currency not present in the trade itself (e.g. when the exchange charges you with their own coin)
+
+### Changed
+- Changed to Python 3.9 (most visible w.r.t. no more imports of List/Dict/Tuple from typing)
+- Streamlined price lookup
+- Streamlined execution normalization
+- Refactored common code and moved some code into more appropriate modules
+
 ## v1.0.2 - 2022-04-01
 
 ### Added